Assignment3_final_version
